snapshot things seem to be working alright
- few thing have changed will list a few
- got rid of page-break-lines wasn't able to use
- @@ -102,6 +99,45 @@ is me seeing if the placement changed anything.
- but did not see anything diff
- @@ -120,7 +156,12 @@ here I added a new keybinding for org-organizer
decided to use a binding that was already in use so needed to change
that one as well
- tried flipping setting in corfu to see what it did...because
there seems some overlap in what ac settings and corfu setting sound
- similar. corfu didnt complete by it's self after reading more it
- sounds like corfu needed a backend as it were to start giving lots of
- candidates, so I found one Ispell, then i'm told ill also need
- auto-complete. ???? so I'm confused by corfu's settings when
they seem to be set by the backend and not by Corfu still undecided
on the overlapping settings.
- @@ -343,13 +341,14 @@ here I fix an indent, and changed
- completion-styes anouther example were I don't see these Corfu
settings change a whole lot of anything. because I changed this to
something other than just orderless and did not notice any different
candidates popup seemed the same so after changing just left it changed.
- @@ -364,6 +363,9 @@ grab this from the github page.
more as a reminder of a setting that can be changed. changes
were made to corfu since I started using it. had to read more to
figure out what that actually was and what it ment. not to much
had to do with what seperated words.
- @@ -473,12 +475,18 @@ here I fill out the rest of the default
config from GitHub, trimmed it when first installed.
- @@ -541,30 +549,30 @@ here I uncomment all that I did before
the keybindings don't seem to work but the other settings are
worth having deal with the keybindings anouther time.
- @@ -580,6 +588,7 @@ this was to try this tab-line or better
name would be buffer-line is just that a tab bar listing buffers
didn't see how it was of any use and it squashed the screen so
purge it.
- @@ -614,6 +623,8 @@ this....this to me is the most important change
this setting solved a huge announce and that was having my cursor
jump to the top header after I commeted or uncommented something.
when I first started using M-; it would just work some were that
changed and it started doing the jumping around thing. by this
time I have a gotten accustom to using it and would forget to
use a diff keybinding this fixed it so I don't have to and can
go back to using it.