latest changes
@@ -55,7 +55,9 @@
this is me learning what what I can do with setting my own
key bindings, did lots of work and learning, just messing with
the dashboard. thought it might be useful to have a keybinding
for it. have not tried this yet just fixed up what I had.
so work still in progress.

@@ -78,7 +80,13 @@
this ...this is what I want to talk about. this was my own idea.
and I was able to make it happen. an elfeed icon that opens rss
from the dashboard. IT WORKED! So this is the story.
I have been using an rss feed for quit a while now mainly news-flash.
worked well. one of those programs you set it then almost forget it.
simply because it does what you want and just keeps on doing it well.
so the idea of switching to elfeed wasn't something I was looking for.
if its not broke why fix it was the mentality. plus elfeed did not look
that appealing. did a quick try did not like it so left it and moved on.
as time past i would come across people like distro tube, mike zamasky,
system crafters, all give good reviews of elfeed. read articles once
again favourable reviews. So thought I would give it anouther shot.
needed something to figit with to try and buffer the 24 hr propaganda
that is beeing forced on us, this is what I chose. So to make a long
story longer. since rss feeds are what I start my day with, usually
one of the first things I run. it would make sense to put quick way of
doing that from the dashboard. after starring at my dashboard it hit me
yes a icon/shortcut that opened elfeed from the dashboard. would make
sense. so after looking at the code I was sure I could do this and
just like that I wrote out what I thought it should look like and
copied it to my config with only one miss step, I was able to get it up
and running so the TL/DR is I'am proud of this cause I did it all myself.
AND IT WORKED! an original idea made to come to life.great!

@@ -345,10 +353,10 @@
this was something that I realised that when I changed what I thought
was just corfu's completions it affected the mini-buffer as well.
took a while to notice the difference that happened. needed to switch
back because I find orderless completions in mini buffer way more
useful than the completions at point.

@@ -495,11 +503,6 @@
this just me moving the old elfeed block to the bottom of the config.

@@ -623,8 +626,15 @@
consists of thee different settings.
the first on the "toggle comment" to me is the most important.
at some point an update or something changed the way M-; commented
lines the new behaviour was when you tried to comment out or comment a
line it would jump the cursor to the top header. super annoying. not bad
if small block but something like org-block that large it would move the
entire page up leaving you unable to see what you had just done. the
problem as far as I could see was the command was actually DWIM which is
do what I mean and it was not just for commenting. started trying to
change my all ready learned key and change it to gc for nerd commenter
then I came accross this in my internet wonderings grabbed it right away
this fixed that and I'am much happier now that I can go back to what I
was already doing. great ..like it.
the "ytdl" is new. not much to say here other than it something i'm
looking into. had liked the fuzzy version of this in the terminal.
so just something left here to remind my self. always looking at my
config file. more than most other files.
the third thing is something at first just thought what ever, this is
just an accessory to make your emacs fancier. then realised that I
could use this and I like it. It is useful not just for looks.
the thing i'm talking about is "beacon" it flashes the cursor
when ever it has jumped showing you were in the buffer it is.

@@ -697,8 +707,45 @@
lots going on here as well...first part is me filling out my
mpv.el package config,left this undone till I could go over it more
thrall. basically expanding thing as I learn them. here I figured out
that one can use spc m l l which will be up a list of types of links.
by choosing mpv and filling in the path to a video you can just click
the link and it opens said video in mpv controlled by an org file your
in. mainly time stamping different places in the video with notes that
you can than just click the time stamp to get there.
the next part was me removing a feeble try at number the completions so
one could just hit the appropriat number to finish the completions.
but after tweaking the Ispell auto complete it not as needed now the
completions seem straight forward, no need for this as of now maybe in
future i'll think of this but if not needed now no use in keeping it
here.
the next chunk is for deft, people on emacs are how can I say.....
fucking nutters about their notes, there is such a vast array of
different ways to take ,keep,search notes. It'll make your head swim
came across deft in the process assumed it was just anouther variant.
fired it up just for shits and giggles. realised what it actually did
which to me is it a stand alone note searcher using key word searches.
the reason I was looking at any of these, was the fact that my org
folder is getting harder to find things. and just by fluke I happened on
to this which is just what I was looking for. Now I can leave all the org-roam
and zettelkasten note taking for anouther time.
fter that comes my new elfeed config block. made my own keybinding.
because I can...installed elfeed goodies as well as elfeed-org.
and added there settings as per default. Was able to update feed with
my feed list from newsflash rss feed reader, then made an org-mode file
for all the feeds that can be used by elfeed this all worked well.