wanted to document this change I made
the change I'm refering to will come after this is
the before....then the real change that I made that
I need a record of it for posterity.
this is were it started from

@@ -769,6 +769,22 @@
here I add nookies elfeed-youtube-downloader.
after writing it to the config I gave it a try.
it worked just like it was suppose to.
ended up downloading 7.0mb video in aprox little over a minute.

@@ -776,9 +792,10 @@
added keybinding cleaned up others.

@@ -993,6 +993,22 @@
and
@@ -1000,9 +1016,10 @@
are the same just in org file...just line numbers change