this is the change I wanted documented
in line # 997 in config.org I changed this from youtube-dl to
yt-dlp, the reason was me snooping around and seen this was what
arch appeared to be using. so changed this one line and no others
after witch I gave it a try and it seem to do the same job just a lot
faster... but we will leave like this and see how it works after some
time. and need to get new feeds in. after the new feeds try again and
see if it still working or not