the next installment of this is my config
@@ -100,8 +100,8 @@
changed this because its annoying in elfeed.
messes up the windows text in elfeed. changed to text mode instead
fixed the problem

@@ -689,7 +689,7 @@
this I just noticed, has been wrong since it inception.
the prefix is most all toggles so fixed this.

@@ -726,7 +726,7 @@
here i've decided to move some of these personal files out of .emacs.d
to .doom.d this way they are under vc and can be replaced if needed.
or kept updated. will move them as I come across them.

@@ -749,16 +749,36 @@
sorry you had to see this! It's a fuckin mess right now.
this is me trying to figure out why none of the bindings don't work
still right in the middle of mucking about.
would seem to me it has to do with the mode, in org mode doesn't seem
to recognize spray mode ...ergo no binding .....this is what i'm trying
to deduce. only a theory right now. will figget with it again.
brings me to this point.... I find just throwing what I want to try
right in the config and mess with it there. not sure what advantage
I get by doing the same in separate buffer and just evaluate it from
there. have tried it and ya ..... it worked. But once I had it I would
still have to transfer it over and do a whole save restart anyways.
probably just me and my old ways. (did this for everything else)

@@ -815,6 +835,10 @@
this was something I noticed this in some configs online.
mainly had the problem myself...telling emacs to play a file rather than
edit. this I thought there must be a default way to do this but no I
could not see one and seen this in configs though this must be it...
seem overkill for something most programs do by default..but what ev's
but before I did all of that I thought to look deeper first and it was
not looking like a default way ...it looked as if it's not maintained
and old. reports of bugs,somelinks were dead or moved. not many stars.
after all that decided to not bother and find a work around which I did.
just use mpv-play it would open it and you can controle with org-doc
so no need for this after all will delete soon. sloppy wasn't ready for
picture, kinda decided just to do it.

@@ -930,7 +954,45 @@
had to add these....figured I would didn't make sense why he'd put them
there if they were not needed.live and learn.

@@ -993,14 +1055,25 @@
this is I found in my searching, like I had said before I use to use
a fzf in the terminal to watch videos with out downloading so this
made sense. it is modified youtube dl. only thing is lbry links
don't work so thats ok that is odysee is for. it appears he just sends
mpv to open the url and mpv takes care of the rest. need a protocol
thou. still good. as of now I have many ways of dealing with video
or any other file or url in my elfeed. lots of options that a bonus.

@@ -1009,17 +1082,46 @@
this is what I said earlier. lots of option when dealing with my feeds.
here I add two more eww(emacs web wowser) and w3m. both are new to me.
but should be able to incorporate spay with one or both not sure.
still learning. but that is why spray is being worked on right now.
after getting all this ready and working it's then i notice sprays
keys don't work and I'm not sure why... so lots going on all over at
once. elfeeds turning into a huge beast, was only two lines before. Now?

@@ -1051,3 +1153,17 @@
this is the other part of the elfeed upgrade. the w3m which needed a
package. still trying this so not much can be said about it.
but the big deal is I got all three of these to work. well done.
plus learned some more just going through the process. and it worked!
that should do it for now