OK snapshot 3.1 looking like it'll compute
will take a another snapshot the after shot if you will.
TL/DR got auto-complete w/Ispell as a backend as well as having
flyspell working along side of it. seems to work. there was talk of
them interfering with each other but I'm not seeing that they seem to
work. It appears they may have put a small delay in flyspell so the
auto-complete can engage. flyspell needs some fine tuning. seems
after adding something to your dictionary it will still highlight it.
as well as jump to it when told to go to the next error. but won't
bing up anything to correct it only the next real word it should have
jumped to which isn't favourable.