Here are the updates to snapshot3
- so I managed to add the package that I wanted to try as well as
including the older package with no real problems. basically I got
ac-Ispell and flyspell to work. feel like they both work as expected.
not 100% sure but it may give the completion at point to auto-complete
and delay flyspell but it works. it worked well enough that I ended up
adding more packages....even thou I was just making sure there was no
problems later on. fixed a bug in spray and learned more HTML in the
time I was messing around. next thing you bang I'm not only grabbing
packages and installing but I decide ya I now need a language sever
running. Needless to say the thing I thought was going to be a problem.
turned out to work first time with only small effort, had I not seen the
reports of bugs I probably had just did it and not noticed anything.
instead of doing it one step at a time.

@@ -119,7 +119,7 @@
this was to see if changing the binding changed anything answer: no

@@ -414,12 +414,12 @@
here I've uncomented vertico's "completion-in-region" had commented it
out to debug a problem with dual pop ups. problem got fixed. so undoing
what I had done and see if there is any change so far seems fine so I'll
leave it.

@@ -500,10 +500,23 @@
here is another long story, if everthing else wasn't enough?
since the last commit I dabbled in tables as well, and here is
my use case scenario. I do tend to look at my config for keybinding.
so here is one of those places. still leaning these.

@@ -740,11 +753,12 @@
this is my favourite part, this is were I took a problem and
instead of using something from the web. I fix my spray problems.
first prob beacon mode would show up in the spray buffer not good.
but I solved it on my own by reading the code I came across setting
that would toggle minor modes off. just what I needed and ya it worked.
The next problem was the cursor started showing up as well, my theme
gives me a pink cursor in spray it was this huge pink box at the end
of every word...not good! the seem to imply the cursors value to be nil.
this should have been enough but regardless the cursor was there and
hard to just ignore. yet I had all ready grabbed a snippet from online
that was there to do this very thing and it was not working. after
looking at this snippet I realized it was really hackey. the commands
and function's didn't make a lot of sense and don't believe they worked
anymore. after looking at vertico's intangible-mode which is to keep
the cursor out of the mini buffer. tried this first and had high hopes
that it would work ....but no it was still there. then after checking
the manual I came across cursor-in-selected-windows and set it to nil.
the command it self doesn't sound right but the manual described it as
the command I wanted, so I gave it a try and poof! cursor gone from
spray, so yay 2 problems fixed by my own means not just copying and
pasting. with that I looked at still yet another problem line numbers
had also started showing up in the buffer, another announce.
then I realized I could include it with the minor modes that got toggled
off when spray started. bingo bango another fix in the bag. Done.
so ya did some real good lisp tinkering and most of it worked.

@@ -812,8 +826,6 @@
note this is what did not make the cut. more worried about flyspell
and what modes it may be on were I would prefer it not...this is work
in progress have got enough one on one time with all the new settings
so have yet to come to any conclusions yet.

@@ -1084,6 +1096,7 @@
added white space for defining chunks which I prefer.

@@ -1095,12 +1108,8 @@
these bindings did not work in elfeed-show-mode so, .. purge!

*note in between commits Emacs has decided to release 28.0.1
so I had to deal with that as well
the last commit before got rudely stopped by me losing the diff.
and not to being able to bring it up but yay on top of all this
is when I decided to learn some HTML....this got me installing
languages as well as lsp-mode and that whole thing with servers and
all, so actually lots of changes in what I thought of as a short time.
but like I said before this config is in constant change.