readme success#10
think I got it to were it was only in org-mode.
this was good around got to flush out somethings in org-mode.
got my read me in one format that I'm familiar with. liking the idea
feel like it make messing with the readme way easier.
I may even change the real file in order to make updates easier.
So mission accomplished. success!