New package added,it's config has been intergraded
Reference config.org
@@ -64,6 +64,9 @@
this was grabbed from "what's new in emac 28.0"
lead to believe this is a better setting for emoji-font

@@ -72,6 +75,10 @@
I've come across this multiple times, on other peoples config.
thought I would try add see if there is a diff. unsure as of yet.

@@ -898,6 +905,10 @@
this new keybinding is really useful once again it was one line in a
article about some of his time saving keybinding, this is the only one
that sounded like something I might use. started using right away and
found it super usefull. this is a keeper.
the other change I think is to keep flycheck from throwing up warnings.
there aways seems to be warnings, even when you don't need them they
pollute the buffer with tons of flags this should lessen that.
hopefully, now should mostly just errors. no warnings.

@@ -1113,6 +1125,28 @@
elfeed is were there is the most changes, first started putting a
keybindings table at the top of the header so when I need to check for
one I don't need to scroll through it all just to find what I need.

@@ -1185,11 +1219,20 @@
this is the result of adding a new package called elfeed-tube.
I first installed and added the config to the bottom of my config.org
got it working then figured that it didn't need it's own heading it was
small enough it could just be added to elfeed, this is the result seems
to work fine.

@@ -1202,8 +1245,8 @@
header fix

@@ -1331,29 +1374,3 @@
this is what I incorporated into elfeed.