vterm added + solid minor changes
vterm added, dependants added, config added.
got up and running, works well and snappy.
*note most went well, reason being, stumbled across Erick Navarro's
config online. I did not pick up from the pkg readme the need to
add this but it got it running
:custom
  (vterm-module-cmake-args "-DUSE_SYSTEM_LIBVTERM=yes")
  (vterm-always-compile-module t))
the always compile is more convenient, since it would
stall and wait for an answer.

@@ -32,6 +32,9 @@
- read on reddit this is handy to have at the top
of your config only, left running it will throw
a error when opening an org-edit-src-code blocks
window. will remove soon. have not needed all this
time, if one day there is a need for it i'll come
look for this record of it.

- other then vterm set up the rest was just tidey things up.