ascetic changes as well as new packages
7 files changed, 568 insertions(+), 391 deletions(-)
thats alot of changes, need a snap shot from here
Readme got an update but is already behide these changes
will update it soon.

@@ -48,71 +48,61 @@
added Iosvkem font but left commented

@@ -315,17 +297,16 @@
corrected the hook to the mode it should have been in

@@ -504,30 +487,37 @@
comment out all the company, and super-capfs see what -difference it
makes as of now it's small and suttel so not sure if I need these at all
or just a redo still working on it.

@@ -836,13 +815,17 @@
+
+;; found in manual for eww w/spc h R ;;;;
+(setq eww-retrieve-command
+     '("brave" "--headless" "--dump-dom"))

@@ -871,19 +854,47 @@
added pkg declutter not sure may keep this as a dependant

@@ -1402,8 +1423,16 @@
added dwim-shell-command pkg
this was a handy little pkg for lots of things from pinging the internet
to cloning a git to the proper folder, lots more have used and approved

these are the ones thought should be mentioned, the rest are ascetics