enough changes warrented update
@@ -2,6 +2,7 @@ ;; try running with out indent, don't believe it's
usefull here
@@ -143,10 +144,17 @@ => added shortcut to the fallback +dashboard for
the new journal module installed
@@ -156,6 +164,10 @@ => set default folder for new journal
@@ -197,6 +209,10 @@ => made personal org-capture-template of A file I
call organizer
@@ -220,13 +236,14 @@ => added an after macro here thought it should
have one, thinking it may help...did not see any big change but left it
on anyways
@@ -332,13 +349,15 @@ => revisted keychordss and 3 new ones, fine tune
delay, may need more tuning
@@ -819,10 +838,19 @@ => this should keep the workstation tab bar
visable more. don't all ways remember that I'm in a seperate workspace.
this should help that may need to change the face, bright-pink little
distracting
@@ -865,8 +893,8 @@ => changed declutter-engine, seem to be expensive
using rdrview, maybe it would get better but as of now not needed. eww
should work for how i've used it
@@ -907,7 +935,7 @@ => this use to work but then stopped. after a little
digging it appeared that the command visual-motion was deprecated and
found state was more widly used. appears to have worked. yay i guess.
@@ -944,7 +972,7 @@ => trying settings to see what I prefer
@@ -1006,6 +1034,7 @@ => was sure this use to work before, but it failed
when I went to use it, prompting me to add this, didn't seem nessesary
before?
@@ -1201,9 +1229,8 @@ Elfeed => this is work in progress, me fuckin with
shit. in hide site I believe it was just paticulare websites.
@@ -1228,6 +1255,7 @@ Elfeed => experimenting with keybindings for
elfeed-show-mode
@@ -1248,10 +1276,45 @@ Elfeed => this is two experiments in one.
first is a hook to start elfeed-summary, every time elfeed is started.
this has worked and I may keep for a while.
the second is a new package that I discovered, the code sound like
something interesting, wanted to see it in action. still working on it.
@@ -1361,6 +1424,9 @@ Elfeed => this is hook #2 chained to the first
hook. the way this should work and does is:
start elfeed => hook#1 starts elfeed-summary => hook#2 updates feed.
this seems to work, well enough. one of the reasons for this is I get to
see the updates progress, know exactly when it has finished. no real
indicator that I could tell in elfeed when it was finished. now I can.
did read that it slows the process down a lot but it's doesn't seem that
bad? so far I consider this a success.
@@ -1460,6 +1526,7 @@ open source map => this is a test. it is a bunch
of predefind commands for the shell-dwim package. not sure if this will
just pollute things or not if so I'll remove it and just run the couple
I have used, yes I have used this package and it worked perfect. so more
is better ....maybe...maybe not.
@@ -1507,3 +1574,4 @@ open source map => this was left here just to
ponder it .... long story .