changes made. need to record
this is not what I would call a snapshot.
this is just a record of were I't is as of now.
for some details doom at the moment has made some updates
that has caused small problems here and there, that get fix fast
had small hiccup with doom s -u . it fixed itself. so things seem to
all be working as expected. but have not been doing much with it since
the hiccup, but understand there are possible breakage happening as
updates are rolled out. this made me think now might not be a bad time
to leave a updated copy of were my config is at so far. so this is it as
of september 11 2022.

7 files changed, 173 insertions(+), 85 deletions(-)
@@ -103,7 +103,7 @@ => this causes the one character short cuts in the
dashboard to not work, this fixed that, probably just get rid of this
line.

@@ -167,6 +167,7 @@ => this changed the journal from daily to yearly
putting it all on one file rather than a new file for everyday.

@@ -210,9 +211,14 @@ => this is one of those thing that will take a wall
of text to explained what happened here. the tldr is added a new capture
template should have been straight forward but ran into problem, fixed
problem, but don't understand what the problem was or how my solution
worked. was the indent....was it the parrends...??? not sure...works
now.

@@ -349,12 +355,12 @@ I'm getting better at typing, need to fine tune
these as not to accidentaly trigger them when not wanting to.

@@ -519,6 +527,15 @@ => work in progress. trying to get better
candidates by hacking together something I found online but threw an
error so hacked it to make it work, no error but can't say it did
anything. Will need to revisit in future and figure out what needs to be
done here,

@@ -793,6 +810,7 @@ => here is the results of that hiccup that happened
when I tried to update. It told me basically that the receipt was wrong.
at this point I had a project on the back burner were it appeard that
one could just host the package on your own system. Fixed the receipt.
then thought here is a reson to implement that project that I had planed.
So here is were I changed this to grab the package from the repo I
downloaded.

@@ -838,6 +856,15 @@ => this has the manual pkg config as well as my
failed attempt to get a simple total line count. the total line count
is still not feesable. just something I try now and again thinking maybe
this time but not this time. it can be done it just seems alot for
something not that important.

@@ -873,10 +903,10 @@ => this one also got an error when reading it
receipt. but this I find hard to find places to use it. Don't run into
walls of text like I use to, most things are in a formate that renders
it n/a

@@ -1279,38 +1309,43 @@ Elfeed elfeed paywall ....no comment.