lots of changes slow in pushing them
don't like doing pushes when there is something that is still be
actively worked on. knowing it either be changed or gone completely.
this prevents me from saving to git regularly. anyway so what, big
whoop. list some changes from config.org that may need it.

@@ -37,15 +37,12 @@ => this is some decluttering, don't believe I need
these, will comment out for now, remove later if not needed.

@@ -82,7 +79,8 @@ => need this for sane numbering on folded files.

@@ -102,12 +100,6 @@ => decluttering this didn't seem useful.removed.

@@ -145,7 +137,9 @@ => learning more about the agenda, thought I would
try this out and see if it is useful or not.

@@ -168,6 +162,7 @@ => I liked this one, added auto-fill-mode to
journals, this keeps lines more readable as well as keeping it looking
neater.

@@ -211,14 +206,18 @@ => started hacking and tweaking
org-capture-templates. more to come here, so this won't stay the same.

@@ -231,10 +230,10 @@ => change src-block editing buffer, getting use to
the split below now that I know how to raise and lower it. so I thought
I would give the default a try. so far so good. the other is the
auto-save-idle-timer, I like beening able to revert buffer back simply
by not saving it, this complicated that. so I commeted-out, possible
remove later.

@@ -259,13 +258,6 @@ => declutter

@@ -292,7 +284,8 @@ => flycheck complained this did not have a group,

@@ -362,8 +358,8 @@ => try improving some keychords.

@@ -496,64 +492,35 @@ => here I am decluttering and trying to inprove
cape completions, so this is work in progress.

@@ -856,15 +823,11 @@ => here I added a trash bin rather than just
deleting. the rest is a failed project.

@@ -943,6 +903,8 @@ -> added the newest trend and that is the
webster-1913-dictionary.

@@ -1047,9 +1009,6 @@ => had read in the updates that this was
deprecated and not needed so I removed it.

@@ -1259,8 +1218,7 @@ Elfeed => after update emacs complained about this
so I simply made it the same as the rest, got rid of the problem and
seems to still work.

@@ -1279,22 +1237,32 @@ Elfeed improve elfeed

@@ -1310,6 +1278,7 @@ Elfeed => here I tried multiple ways of making
opening elfeed-summary and elfeed-update automatic they all worked but
in the end I decided to go back to the default for now.

@@ -1349,6 +1318,9 @@ Elfeed => this is something that I learned and
what to make a comment about, learned that if you want just one header
to always be folded, you can use properties setting to do this, so now I
don't have to navigate up and down elfeed-summary-settings, they are
longer than most, evertime I scroll my config.

@@ -1531,6 +1496,14 @@ => I find these dwim fuctions irresistible