multiple updates and changes
8 files changed, 869 insertions(+), 542 deletions(-)
diary file added, not sure read it in info doc, so added
run down of changes to config.org

@@ -1,26 +1,30 @@ => added a preamble type drawer for aesthetics,

@@ -29,20 +33,24 @@ => toggle debugger in menu, so removed for non-use
- moved add-load-path up into package manager section, this just makes
more sense

@@ -59,7 +67,7 @@ package management => bold looks better

@@ -78,6 +86,7 @@ package management => have not declared this anywere
so thought it couldn't hurt to have it declared.

@@ -95,103 +104,135 @@ => *note: realized putting the link to the github
page of the package that the headline refers to only makes sense.Will
start filling more as needed.
- last update uncentered somethings thought of move command up the chain
would help but no
- next replaced opening browser to opening agenda on the "nav buttons"
- As well as one to add journal entry
- needed to reorder thing since emacs stoped resetting the cursor after
a doom sync needed "restart" at the end for one extra tab to get there
- removed some old setting not being used
- added elfeed to the +doom/dashboard or fallback dashboard, more useful
then what is already there.
- while looking into how to add too the menu. Came across a deprecated
command. So changed that as well
 * ;;;; org-settings ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 - git has decided to clump this all together so I'll put the divider

- first added link under header to org-manual just makes sense
- changed diary-file should not have been an .org file put in vc folder
- the next several sections are the tweaking and upgrading of the
journal-file, everything from capture templates to agenda interacting,
as well as used the customize-ui to add minor-modes like auto-fill-mode
needed to change the carry over so it came be used to schedule things
with a to-do and time stamp all from the comfort of a
org-capture-template. So it is upgraded to be more useful
- after that is a clean up of some org keybindings, less clutter

@@ -209,20 +250,20 @@  => here we are learning I think. getting more
familiar with the customise-ui and did more reading. I was told not to
abuse the add-to-list command, and realizes that they were being stored
in the costume-file, to add to everything. the org-capture-templates
section is getting huge, and remember trying to put the whole thing in
config. Emacs complained hard and was super slow in starting.
so before adding more, removed some older ones. there still there just
not were I see them all the time.
- so added a new one to schedule app. in the journal and be included in
agenda
- next one I came across in my research, for keeping bookmarks of web
sites

@@ -234,54 +275,54 @@ => this mostly clean up and aesthetics

@@ -320,8 +361,9 @@ => adding links to relevant info
@@ -333,6 +375,7 @@
@@ -352,6 +395,7 @@

@@ -363,14 +407,16 @@ => update keychords add new ones remove others

@@ -433,13 +480,13 @@ => corfu revisted, updated, straighten out

@@ -485,15 +532,20 @@ => added corfu-history to the mix, to make
completions more efficient. In theory, will see.

@@ -516,10 +568,13 @@ => clean up

@@ -557,6 +612,8 @@ add links
@@ -580,6 +637,7 @@

@@ -784,22 +842,22 @@ change to emacs to manage this package, fixed
parend.

@@ -829,10 +887,20 @@ => add function for commenting and uncommenting

@@ -916,27 +983,53 @@ => add language-tool prose-linter or grammar
checker, add keybindings, could not get intergraded in to flycheck.

@@ -1049,93 +1142,112 @@ => MPV work-in-progress
have been monkeying around with this for along time, was not getting
what what I wanted and I thought should be easy.... multiple revisions, still not
finalized

@@ -1188,14 +1300,14 @@ Elfeed => got rid of goodies (deprecated)
noticed it was the thing giving the "cl deprecated error" at start up

@@ -1229,10 +1342,27 @@ Elfeed => added two functions
- one is declutter
- this was to try and use in elfeed-show-mode ?? not sure

@@ -1589,9 +1723,26 @@ => added function to grab youtube-subtitles