project big update #1 completions
concluded that I'm doing it all wrong,
- come to the realization that I am double loading and double configuring
most packages.
- kinda knew there was some, but then I came across doom's config.el
files. then I realized the extent of it.
- my plan is to revert these config to doom default then work from there
- lots of changes to come will do in stages
- archive old settings
this is going to be a large change
some things I learned were dooms config files were at, and how to comment out
a src-block using the org-header this fixed the ":tangle no" being missed
or forgotten. TLDR using "C-c ;" adds the key-word "comment" to the
header, src-blocks will be included if inheritance is "T". this makes it
easy to do as well as see it.

@@ -728,110 +730,150 @@ consult
- It started with consult so here the first attempt at it
- plan is to archive old
- add in doom's with the ":tangle no" flag, then in the future I can use
ether of them also tag header-title with doom so I know it not tangled

@@ -1118,12 +1080,20 @@ org-rich-yank
- going to this out, wraps snippets in blocks. we will see?

@@ -1175,6 +1145,7 @@ evil-snipe
- this is part of previous test, tring to get diff header names
still working on it

@@ -1187,8 +1158,6 @@ easy-motion
- not exactly what I wanted, found par-edit, shouldn't need this.