changes were made....repos need updating
still in the process of integrating my config with doom better.
added more pkgs cause ....why not?
here some changes that maybe need more description
this will be from config.org

@@ -112,6 +112,11 @@ Startup-settings
- added the save place setting into start-up dont plan on changing them
so no need for it's own header

@@ -170,7 +175,7 @@ dashboard
- this don't look like much but it is predefinded by doom and does
couple of things one of which is it open it in it's own workspace and
names it rss. Like this part, however it opens it to the right of the
main page and I have got in the habit of opening another workspace to
the right. this causes me to never use that first workstation not a big
deal. keeping this for a while. I think there is a command to move said
workspace and probably a easy way to add it to the config and have it
automated ....what ever!

@@ -299,13 +304,13 @@ org-settings
- like this package "org-pandoc-import" just never sure if I'm accually
using it or got it set right. have the ability to do it with pandoc but
like how this makes that even easier. So did a quick check and updated
it to the doom-way and added some better description

@@ -317,7 +322,7 @@ org-settings
- reset after debugging

@@ -440,6 +445,90 @@ doom-Smartparens
- added this for reference :tangle no (already is)

@@ -453,27 +542,52 @@ evil-surround
- adding doom-mod-configs and reordering some headers

@@ -711,7 +830,7 @@ cape
- comment by minad on a forum made me take a closer look at my settings
try some fine-tuning. the TLDR is he proposes that triggering them
manually is better.

@@ -939,38 +1050,7 @@ consult-doom
- more doom intergation and references

@@ -1267,12 +1375,13 @@ personal-settings
- one of the typing exercise packages that I'm tring out

@@ -1287,12 +1396,13 @@ personal-settings
- new function for occur to put cursor in it's buffer after search

@@ -1301,14 +1411,14 @@ way of pasting that automatically surrounds the snippet in blocks,
- add some links here. this sounds like something I would use but have
not used much mainly because I forget its here. added a not to remember
next time.

@@ -1346,26 +1456,204 @@ my-keybindings
- added doom-keybinding reference

@@ -1867,7 +2137,12 @@ Elfeed
- BUG after reverting emacs back to 28.2.2 something got lost and there
was no function definition for elfeed-expose. found it and just added it
it was small so hacky but worked. did work so thinking this is temporary

@@ -2102,8 +2380,8 @@ EWW
@@ -2150,9 +2428,9 @@ EWW
- trying to debug why eww chokes on pdfs

@@ -2456,17 +2746,98 @@ logos
- try another "Prot" package this time logos
- the problem is I only get about 30% of Prot's packages to work. not
sure why if it's just me or the translation to doom not working not sure
but like the other Prot packages I have yet to get this to work properly
- will keep "f"ing with it for now.