freshly picked,new updates,and new and improved
- integrating dooms way of doing things is already improving things
- still on going. plan on doing chucks at a time rather than change
everything then try debugging after time chasing other bugs forget what
the cause was. So changed bttm ten headers. then made sure they seem to
operate the same.
- added new package, allows for searching of browser-history.
list of changes that may need more details
config.org

@@ -312,6 +312,11 @@ org-settings
- long story, simple idea, still sorting through it but the tldr is I
like the idea of changing as many docs that I want to read to org.
simply and easy. Seem to have got confused with importing and exporting
then throw in pandoc, and a seperate package that all has to do with
similar things. got confused. left it. Now back to tring to straight
some it out, made some minor change to keybindings. added an export.

@@ -780,7 +785,7 @@ corfu
- this a test to see what difference this setting makes, try the
defaults. So far not much of a change. will keep accessing for now

@@ -820,11 +825,24 @@ cape
- in my config I have links to the read-mes and github pages of the
package, did not think they needed to be in the read-me ...not sure.

here we are doomifying the bttm ten headers

@@ -2514,9 +2532,12 @@ dwim-shell-command
@@ -2535,7 +2555,8 @@ v-term
@@ -2573,6 +2594,7 @@ engine-mode
@@ -2604,7 +2626,8 @@ yt-sub-ext
@@ -2622,7 +2645,7 @@ markdown-mode
@@ -2660,7 +2683,9 @@ language-tool
@@ -2707,7 +2732,7 @@ Denote
@@ -2723,7 +2748,7 @@ yeetube
- finaly at the bttm I added the new package
had a rough time getting it to work...long story
- the package should allow one to search browser-history in a buffer
then go right to it, helps if you did bookmark place but remember
visiting.
@@ -2842,12 +2869,30 @@ hl-to-do-mode