small and big changes. still work in progress
@@ -40,27 +39,18 @@
- what I could decipher was one deosn't add packages to path as you
are told by others. the way it sounds, you add the path in package.el
then use "require" or "use-package" to declare and load it. no need to
add-to-path. So had to do some
commit,restart,delet-old,add-new,resync,fix-errors,resync,restart.
so a lot of fucking around and all I have to show for it is this diff
and a better Emacs, so fuck it, PURGE IT ALL!
- that little (add-to-path! "+config") took me all day to get
operational. It was my attempt at making a "features-file" to make a
nav-button to open the config file in it's own named workspace, believe
I was able to do it using dooms =rss as a template. then put it on the
back burner for now. will revisit later but it's mostly do and this is
the setting to use it.
- added a new directory "myrepo" were I will start to add my
manual-builds and packages.

@@ -122,8 +112,6 @@ startup-settings
- gab this from mastering-emacs post and on a reread it appears to
be for mac's, so not much use to me .... purge

@@ -136,9 +124,11 @@ startup-settings
- did not know how to wall one function from another in
"lisp-interaction-mode" so change scratch-buffer-mode to org and use
src-blocks, this is the way I'm running the configuration so why not the
scratch as well, will see if this works or not time will tell.

@@ -207,6 +197,26 @@ dashboard
- reinstalled my +doom-dashboard-menu-sections like them there so
brought them back, still working on the best way to manipulate this menu
still in the trial and error phase, this was just anouther step in that.
tried remove from list some sections but they reappeared, now I have
resorted to doom's modular-config files though these would be over
written but was surpised it did not ....so more experiments to come but
get these back for now. still have a menu item commented out and it
seems to have stayed. so more to work on but works good now.

@@ -250,12 +260,19 @@ org-settings
- added "org-ros" this look like a convenient way of adding in-line
images to org-files so gab it up but have not got around to trying it
yet. ???
- the other thing going on here is changing org-capture and org-tempo to
use-package and doomify this section, had problem with tempo, reverted
it back only to relies, I had been trying to use wrong keybinding and
that was why I believed it was not working, will revisit anouther time
to much time spent on it at this point, can wait for another time.

@@ -266,33 +283,32 @@ org-settings
- added the after macro and unwrap the prefix section in my map! macros
they don't seem to be a thing, did not see anything about it but, get
warnings from flycheck and when I go look for references in doom
discourse, none of them are wrapped so I'm assuming that it has been
depreciated? seems to be better without so removing as I come across
them.

@@ -810,7 +826,7 @@ cape
- attempting to fine tune cape.

@@ -1163,18 +1171,18 @@ consult-doom
- brain fart here thought for a moment that m stood for map but it is
real motion and I knew that before and forgot.???

@@ -2151,36 +2130,59 @@ Elfeed
- here added elfeed-curate sounds interesting still need to try out.